Fix oauth with streamable mcp servers
req.Clone makes a shallow copy of the body so the first call would
consume the body completely, making the next call after oauth flow an
empty request, this would give us a 400 Bad Request.

Keep the body so that we can re-inject it into the original request
before re-sending with the token.

Fixes #485

Signed-off-by: Djordje Lukic <djordje.lukic@docker.com>